% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-age-specific-cases-by-age.R
\name{get_cases_age_england}
\alias{get_cases_age_england}
\title{Get cases by age}
\usage{
get_cases_age_england()
}
\description{
This function allows you extract age-specific case counts and rates in 5 year age bands as a timeseries.
}
\examples{
get_cases_age_england()
}
\keyword{cases}
