% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-age-specific-cases-by-age.R
\name{plot_cases_age_england}
\alias{plot_cases_age_england}
\title{Plot case-rates by age}
\usage{
plot_cases_age_england(age = "60+")
}
\description{
This function allows you to plot age-specific case counts and rates for England in 5 year age bands as a timeseries.
}
\examples{
plot_cases_age_england()
}
\keyword{cases}
